refactor: split amqp mod down
