+++ src/amqp/definitions.rs
@@ -1,8 +1,8 @@
-use tracing::info;
+use tracing::debug;
 use crate::amqp::Definitions;
 
 pub async fn get_definitions(url: &str) -> Result<Definitions, Box<dyn std::error::Error>> {
-	info!("Requesting definitions for new bindings");
+	debug!("Requesting definitions for new bindings");
 	let res = reqwest::get(url).await?;
 	let res: Definitions = res.json().await?;
